The system will not allow me to select a new/different Check out date if I click on the calendar icon.  I have to manually enter the check-out date.
last commit text- If by mistake the user misses to change the checkout date and checks in the guest and comes back to change the check-out date, the Room No. dropdown blanks out and that particular room becomes unavailable for me to select, so basically once I've checked in a customer, I'm not able to change the checkout date for that customer/room.  I should be allowed to change the date and Time if necessary.